requirements.txt
pillow==11.2.1
pytesseract==0.3.13
selenium==4.33.0
